refactor(mobile): clean up Send localization logic
