+++ apps/mobile/src/components/amount-field/amount-field.utils.ts
@@ -1,37 +1,8 @@
-import { decimalSeparator } from '@/features/send/temporary-constants';
 import { InputCurrencyMode } from '@/utils/types';
 
 import { currencyDecimalsMap } from '@leather.io/constants';
 import { type Currency, type MarketData, type Money } from '@leather.io/models';
 import { baseCurrencyAmountInQuote, createMoneyFromDecimal } from '@leather.io/utils';
-
-interface FormatPrimaryValueParams {
-  value: string;
-  currency: Currency;
-  showCurrency: boolean;
-  locale: string;
-}
-
-export function formatPrimaryValue({
-  value,
-  currency,
-  showCurrency,
-  locale,
-}: FormatPrimaryValueParams) {
-  const decimalPart = value.split(decimalSeparator)[1];
-  const fractionDigits = decimalPart?.length ?? 0;
-
-  const formatter = new Intl.NumberFormat(locale, {
-    style: showCurrency ? 'currency' : 'decimal',
-    currency,
-    minimumFractionDigits: fractionDigits,
-    maximumFractionDigits: fractionDigits,
-  });
-
-  const formattedValue = formatter.format(Number(value));
-  // Ensure trailing decimal separator, as primary value is a live input,
-  return value.endsWith(decimalSeparator) ? formattedValue + decimalSeparator : formattedValue;
-}
 
 interface FormatSecondaryValueParams {
   value: string;
