+++ apps/mobile/src/features/send/constants.ts
@@ -1,0 +1,2 @@
+// TODO: Retrieve dynamically once new formatter is in ready: https://linear.app/leather-io/project/standardize-balances-formatting-e8bf0300f9b5
+export const locale = 'en-US';
