--- apps/mobile/src/features/send/temporary-constants.ts
@@ -1,3 +1,0 @@
-// TODO: Retrieve dynamically: https://linear.app/leather-io/issue/LEA-2375/enable-full-localization-in-the-send-form
-export const locale = 'en-US';
-export const decimalSeparator = '.';
